+++ libs/librrgraph/src/base/rr_graph_storage.h
@@ -92,9 +92,9 @@ static_assert(alignof(t_rr_node_data) == 16, "Check t_rr_node_data size");
  *           Needed to properly draw.                                        */
 struct t_rr_node_ptc_data {
     union {
-        int16_t pin_num;
-        int16_t track_num;
-        int16_t class_num;
+        int pin_num;
+        int track_num;
+        int class_num;
     } ptc_;
 };
 
@@ -468,10 +468,10 @@ class t_rr_graph_storage {
                  const vtr::vector<RRNodeId, RRNodeId>& inverse_order);
 
     /* PTC set methods */
-    void set_node_ptc_num(RRNodeId id, short);
-    void set_node_pin_num(RRNodeId id, short);   //Same as set_ptc_num() by checks type() is consistent
-    void set_node_track_num(RRNodeId id, short); //Same as set_ptc_num() by checks type() is consistent
-    void set_node_class_num(RRNodeId id, short); //Same as set_ptc_num() by checks type() is consistent
+    void set_node_ptc_num(RRNodeId id, int);
+    void set_node_pin_num(RRNodeId id, int);   //Same as set_ptc_num() by checks type() is consistent
+    void set_node_track_num(RRNodeId id, int); //Same as set_ptc_num() by checks type() is consistent
+    void set_node_class_num(RRNodeId id, int); //Same as set_ptc_num() by checks type() is consistent
 
     void set_node_type(RRNodeId id, t_rr_type new_type);
     void set_node_coordinates(RRNodeId id, short x1, short y1, short x2, short y2);
